pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "habit_tracker"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
